Crate pats

Parcel Ascent Tracing System (PATS) is the numerical model for convective parcel ascent simulation in three-dimension developed as a Master’s project at the School of Environmental Sciences at the University of East Anglia.

The model predicts how convective air parcel behaves in conditions much more realistic then those assumed by classical parcel theory.

The model is currently under development.

Modules

constants

Module containing constants used by the model.

errors

Module with error definitions for all struct and function in the model.

model

Module containing the actual model code. Whole documentation of how the model works is provided here.

Statics

ALLOCATOR

Global allocator used by the model.

Functions

main

The main program function. Prepares the runtime environment and calls the model::main.

Type Definitions

Float

Type used by the model floating point variables.