Module pats::model::environment

Module responsible for reading and storing boundary conditions (environment) data, and providing that data to parcels.

Modules

accesser

Module with methods for accessing the environment and surface boundary conditions data.

bisection

Module containg methods for conducting binary search (bisection) of elements closests to searched values in datasets.

buffer

Module containing methods responsible for buffering data from GRIB input.

interpolation

Module containing interpolation methods.

projection

Module with methods to do computations of geographical projection used by the model. Closely follows algorithms and instructions in: https://pubs.er.usgs.gov/publication/pp1395

Structs

Environment

Environment main struct storing and providing boundary condition (environment) data.

Fields

Struct for storing environmental variables from levels above ground (currently pressure levels).

Surface

Struct for storing environmental variables at/near surface.

Functions

approx_central_lon

Function to approximate the longitude of domain centre on the WGS84 ellipsoid.

compute_domain_extent

Function to get a lat-lon extent of domain with margins.

compute_top_lat

Function to compute the latitude of domain top on the WGS84 ellipsoid.

generate_domain_projection

Function to create a geographic projection struct with parameters that allow for lowest distorion for a given domain.

list_levels

Function to get the list of unique levels of specified type in the provided GRIB files.

measure_domain_sides

Function to get domain sides length in meters.

Type Definitions

LonLat

Convenience type to store lat-lon coordinates.