Module pats::model::environment::fields

Sub-module responsible for handling pressure level data buffering.

Structs

Fields

Struct for storing environmental variables from levels above ground (currently pressure levels).

Functions

assign_fields

Reads variables on pressure levels from GRIB file and buffers them in domain + margins extent.

cast_lonlat_fields_coords

Buffers longitudes and latitudes of pressure level data gridpoints.

collect

(TODO: What it is)

compute_virtual_temperature

Computes and buffers additional pressure level data from values previously read from the GRIB file.

construct_fields

Function to read pressure level data from GRIB input in extent covering domain and margins and buffer it.

list_levels

Function to get the list of unique levels of specified type in the provided GRIB files.

messages_to_array

Collects data from GRIB messages on specified level type into a 3d array,

read_raw_field

Reads all values in GRIB file at specified level type of variable with given short_name and collects them into a 3d array.

read_raw_messages

Filters and read all GRIB messages that contain variable with given short_name on specified level type.

read_truncated_pressure

Creates a 3d array of pressure data of shape identical to other pressure level fields.

truncate_field_to_extent

Truncates data on specified level type from GRIB file to cover only the message + margins extent.