Function pats::model::environment::fields::construct_fields

fn construct_fields(
    input: &Input, 
    data: &[KeyedMessage], 
    domain_edges: DomainExtent<usize>
) -> Result<Fields, EnvironmentError>

Function to read pressure level data from GRIB input in extent covering domain and margins and buffer it.

Data from GRIB files can only be read in a whole GRIB domain, therefore it needs to be truncated.

Some useful variables are not provided by most NWP models, therefore they need to be computed.