Struct pats::model::environment::Environment

pub struct Environment {
    fields: Fields,
    surface: Surface,
    input: Input,
    pub projection: LambertConicConformal,
    levels: Vec<i64>,
}

Environment main struct storing and providing boundary condition (environment) data.

Use of a separate struct for handling boundary conditions allows to have a clean API for other model parts to use.

Fields
fields: Fieldssurface: Surfaceinput: Inputprojection: LambertConicConformallevels: Vec<i64>
Implementations

impl Environment

pub fn get_surface_value(
&self,
x: f32,
y: f32,
field: SurfaceFields
) -> Result<f32, EnvironmentError>

Function to get interpolated value of given surface field at given (cartographic) coordinates.

pub fn get_field_value(
&self,
x: f32,
y: f32,
z: f32,
field: EnvFields
) -> Result<f32, EnvironmentError>

Function to get interpolated value of given environment field at given (cartographic) coordinates.

impl Environment

pub(in model::environment) fn buffer_fields(
&mut self,
west_south: (f32, f32),
east_north: (f32, f32)
) -> Result<(), EnvironmentError>

Function to read pressure level data from GRIB input in extent covering domain and margins and buffer it.

Data from GRIB files can only be read in a whole GRIB domain, therefore it needs to be truncated.

Some useful variables are not provided by most NWP models, therefore they need to be computed.

fn assign_raw_fields(
&mut self,
edge_lons: (usize, usize),
edge_lats: (usize, usize)
) -> Result<(), InputError>

Reads variables on pressure levels from GRIB file and buffers them in domain + margins extent.

fn read_raw_field(
&self,
short_name: &str,
shape: (usize, usize)
) -> Result<Array3<f32>, InputError>

Reads all values in GRIB file at specified level type of variable with given short_name and collects them into a 3d array.

In GRIB files data on each level is stored as a separate message. Those need to be collected and only then can be converted into a 3d array.

fn read_raw_messages(
&self,
short_name: &str
) -> Result<Vec<KeyedMessage>, InputError>

Filters and read all GRIB messages that contain variable with given short_name on specified level type.

fn read_truncated_pressure(
&self,
edge_lons: (usize, usize),
edge_lats: (usize, usize)
) -> Array3<f32>

Creates a 3d array of pressure data of shape identical to other pressure level fields.

When data in GRIB file is provided on pressure levels the information about pressure at each level is only stored in message metadata. Thus, information what pressure levels are available has to be extracted and then casted to the 3d array expected by the accesser.

fn compute_intermediate_fields(&mut self)

Computes and buffers additional pressure level data from values previously read from the GRIB file.

fn cast_lonlat_fields_coords(
&mut self,
distinct_lonlats: &(Vec<f32>, Vec<f32>),
edge_lons: (usize, usize),
edge_lats: (usize, usize)
)

Buffers longitudes and latitudes of pressure level data gridpoints.

impl Environment

pub(in model::environment) fn buffer_surface(
&mut self,
west_south: (f32, f32),
east_north: (f32, f32)
) -> Result<(), EnvironmentError>

Function to read surface data from GRIB input in extent covering domain and margins and buffer it.

Data from GRIB files can only be read in a whole GRIB domain, therefore it needs to be truncated.

Some useful surface variables are not provided by most NWP models, therefore they need to be computed.

fn assign_raw_surfaces(
&mut self,
edge_lons: (usize, usize),
edge_lats: (usize, usize)
) -> Result<(), InputError>

Reads variables on surface level from GRIB file and buffers them in domain + margins extent.

fn read_raw_surface(
&self,
short_name: &str,
shape: (usize, usize)
) -> Result<Array2<f32>, InputError>

Reads all values in GRIB file at surface level of variable with given short_name.

fn compute_intermediate_surfaces(&mut self)

Computes and buffers additional surface data from values previously read from the GRIB file.

fn cast_lonlat_surface_coords(
&mut self,
distinct_lonlats: &(Vec<f32>, Vec<f32>),
edge_lons: (usize, usize),
edge_lats: (usize, usize)
)

Buffers longitudes and latitudes of surface data gridpoints.

impl Environment

fn read_distinct_latlons(&self) -> Result<(Vec<f32>, Vec<f32>), InputError>

Function to read distinct longitudes and latitudes in input GRIB files.

fn read_input_shape(&self) -> Result<(usize, usize), InputError>

Function to read a grid shape of input GRIB files.

impl Environment

pub fn new(config: &Config) -> Result<Self, EnvironmentError>

Environment struct constructor responsible for reading GRIB files and buffering data in domain extent.