Crate pats

Technical documentation of Parcel Ascent Tracing System (PATS) - the numerical model for convective parcel ascent simulation in three-dimensions.

This documentation provides a description of functions and structures used in the model. Its main purpose is to make it easier to maintain and contribute to the project codebase. However, it can be also useful for users who want to understand the model in more detail.

Modules

constants

Module containing constants used by the model.

errors

Module with error definitions for all struct and function in the model.

model

Module containing the actual model code. Whole documentation of how the model works is provided here.

Statics

ALLOCATOR

Global allocator used by the model.

Functions

main

The main program function. Prepares the runtime environment and calls the model::main.

Type Definitions

Float